Extract HasChars module
